pyproject.toml
[tool.poetry]
name = "storage-api-proxy"
version = "1.0.0"
description = "A proxy service for executing SQL queries in Keboola Storage API workspaces"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "storage_api_proxy", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
snowflake-connector-python = "^3.6.0"
aiosqlite = "^0.19.0"
httpx = "^0.26.0"
aiohttp = "^3.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 